LayoutTests/sputnik/Conformance/12_Statement/12.6_Iteration_Statements/12.6.4_The_for_in_Statement/S12.6.4_A15-expected.txt

CONSOLE MESSAGE: line 79: SyntaxError: Unexpected token '}'
S12.6.4_A15

PASS Expected parsing failure

TEST COMPLETE
